+++ apps/overmind.sh
@@ -12,14 +12,20 @@ case $ARCH in
 esac
 
 # prepare the download URL
+CURRENT_VERSION="v$(overmind -v | sed -e 's/.*version \([^"]*\).*/\1/')"
 GITHUB_LATEST_VERSION=$(curl -sL -H 'Accept: application/json' https://github.com/DarthSim/overmind/releases/latest | sed -e 's/.*"tag_name":"\([^"]*\)".*/\1/')
 GITHUB_FILE="overmind-${GITHUB_LATEST_VERSION}-$(uname -s | sed -e 's/\(.*\)/\L\1/')-${ARCH}.gz"
 GITHUB_URL="https://github.com/DarthSim/overmind/releases/download/${GITHUB_LATEST_VERSION}/${GITHUB_FILE}"
 
 # install/update the local binary
-mkdir tmp
-curl -sL -o tmp/overmind.gz $GITHUB_URL
-gunzip tmp/overmind.gz
-install -Dm 755 tmp/overmind -t "$DIR"
-echo "[*] overmind (${GITHUB_LATEST_VERSION})"
-rm -rf tmp
+if [ $CURRENT_VERSION != $GITHUB_LATEST_VERSION ]
+then
+  mkdir tmp
+  curl -sL -o tmp/overmind.gz $GITHUB_URL
+  gunzip tmp/overmind.gz
+  install -Dm 755 tmp/overmind -t "$DIR"
+  echo "[*] overmind (${CURRENT_VERSION} -> ${GITHUB_LATEST_VERSION})"
+  rm -rf tmp
+else
+  echo "[-] overmind (${CURRENT_VERSION} - latest)"
+fi
